fix deleting not own account

+++ ItPlanet.Web/Controllers/AccountsController.cs
@@ -60,6 +60,11 @@ public class AccountsController : PublicControllerBase
     {
         try
         {
+            var (email, _) = Request.ExtractUserData();
+            var user = await _accountService.GetAccountAsync(accountId);
+            if (user.Email != email)
+                return Forbid();
+            
             await _accountService.RemoveAccountAsync(accountId).ConfigureAwait(false);
             return Ok();
         }
